update the data type of shipping cost and vat percentage to be float instead of integer

+++ starter/database/migrations/2022_04_30_0515000_create_store_settings_table.php
@@ -18,8 +18,8 @@ class CreateStoreSettingsTable extends Migration
             $table->foreignId('store_id')->references('id')
                 ->on('stores')->onDelete('cascade');
             $table->boolean('vat_included')->default(true);
-            $table->integer('vat_percentage')->nullable();
-            $table->integer('shipping_cost')->nullable();
+            $table->float('vat_percentage')->nullable();
+            $table->float('shipping_cost')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
